Xcode decompiler.
There's a legitimate reason for this being added:

We do not currently have a solution for people who want to run cromwell/xbox linux only
via a modchip on 1.6/1.6b xboxes.

So:

They can use this program to extract the xcodes from their onboard bios chip (on 1.6, you
can turn the modchip off and read the onboard bios, as both are LPC), and use this to
build a cromwell image for their system, as they are allowed to make use of (but not distribute)
the xcodes that came with their system.

I'd like to think that eventually, we'll come up with a set of our own xcodes, but
as we haven't so far, this is at least a stop gap solution to the problem.

I hope to script this shortly.